chore: rename mopeka_types to models
